+++ src/plone/restapi/services/users/add.py
@@ -5,10 +5,12 @@ from AccessControl import getSecurityManager
 from plone.restapi.deserializer import json_body
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.services import Service
+from Products.CMFPlone.utils import getFSVersionTuple
 from Products.CMFCore.utils import getToolByName
 from Products.CMFCore.permissions import AddPortalMember
 from Products.CMFCore.permissions import SetOwnPassword
 from zope.component import getAdapter
+from zope.component import getMultiAdapter
 from zope.component import queryMultiAdapter
 from zope.component.hooks import getSite
 from zope.interface import alsoProvides
@@ -30,6 +32,8 @@ except ImportError:   # pragma: no cover
     # Plone 5.0 and earlier
     from Products.PasswordResetTool.PasswordResetTool import ExpiredRequestError  # noqa
     from Products.PasswordResetTool.PasswordResetTool import InvalidRequestError  # noqa
+
+PLONE5 = getFSVersionTuple()[0] >= 5
 
 
 class UsersPost(Service):
@@ -150,12 +154,35 @@ class UsersPost(Service):
         send_password_reset = data.pop('sendPasswordReset', None)
         properties = data
 
-        # set username based on the login settings (username or email)
-        if security.use_email_as_login:
-            username = email
-            properties['username'] = email
+        if PLONE5:
+            # We are improving the way the userid/login_name is generated using
+            # Plone's plone.app.users utilities directly. Plone 4 lacks of the
+            # login_name one, so we leave it as it is, improving the Plone 5
+            # story
+            user_id_login_name_data = {
+                'username': username,
+                'email': email,
+                'fullname': data.get('fullname', '')
+            }
+
+            register_view = getMultiAdapter(
+                (self.context, self.request), name='register')
+
+            register_view.generate_user_id(user_id_login_name_data)
+            register_view.generate_login_name(user_id_login_name_data)
+
+            user_id = user_id_login_name_data.get('user_id', data.get('username')) # noqa
+            login_name = user_id_login_name_data.get('login_name', data.get('username'))  # noqa
+
+            username = user_id
+            properties['username'] = user_id
         else:
-            properties['username'] = username
+            # set username based on the login settings (username or email)
+            if security.use_email_as_login:
+                username = email
+                properties['username'] = email
+            else:
+                properties['username'] = username
 
         properties['email'] = email
 
@@ -177,6 +204,16 @@ class UsersPost(Service):
             return dict(error=dict(
                 type='MissingParameterError',
                 message=str(e.message)))
+
+        if PLONE5:
+            # After user creation, we have to fix the login_name if it differs.
+            # That happens when the email login is enabled and we are using
+            # UUID as user ID security settings.
+            if user_id != login_name:
+                # The user id differs from the login name.  Set the login
+                # name correctly.
+                pas = getToolByName(self.context, 'acl_users')
+                pas.updateLoginName(user_id, login_name)
 
         if send_password_reset:
             registration.registeredNotify(username)