+++ src/plone/restapi/tests/test_services_users.py
@@ -292,6 +292,28 @@ class TestUsersEndpoint(unittest.TestCase):
         fields = [x['field'] for x in errors]
         self.assertEqual(['roles'], fields)
 
+    def test_add_user_with_uuid_as_userid_enabled(self):
+        # enable use_email_as_login
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.use_email_as_login = True
+        security_settings.use_uuid_as_userid = True
+        transaction.commit()
+        response = self.api_session.post(
+            '/@users',
+            json={
+                "email": "howard.zinn@example.com",
+                "password": "secret"
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+        user_id = response.json()['id']
+        user = api.user.get(userid=user_id)
+        self.assertTrue(user)
+        self.assertEqual('howard.zinn@example.com', user.getUserName())
+        self.assertEqual('howard.zinn@example.com', user.getProperty('email'))
+
     def test_get_user(self):
         response = self.api_session.get('/@users/noam')